+++ lib/src/voyager_argument.dart
@@ -1,6 +1,6 @@
 /// wrapper class for argument passed via navigator
 /// or in the constructor of `VoyagerWidget`
 class VoyagerArgument {
-  final dynamic argument;
-  VoyagerArgument(this.argument);
+  final dynamic value;
+  VoyagerArgument(this.value);
 }