+++ lib/voyager.dart
@@ -7,7 +7,7 @@ export './src/router_plugin.dart';
 export './src/utils.dart';
 export './src/voyager_argument.dart';
 export './src/voyager_widget.dart';
-export './src/voyager_stateful_widget.dart';
+export './src/voyager_stateless_widget.dart';
 export './src/voyager.dart';
 
 export './src/plugins/type_plugin.dart';